requirement_agent.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.34.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter-client==8.6.0
jupyter-core==5.5.0
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langchainhub==0.1.14
langsmith==0.0.71
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.4.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pvporcupine==3.0.1
PyAudio==0.2.14
pydantic==2.5.2
pydantic-core==2.14.5
pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-requests==2.31.0.10
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==2.1.0
wcwidth==0.2.12
yarl==1.9.4
zipp==3.17.0
